unit column moved to values_history table
